[MOGG][Extensibility] Use CreateEmptyTensor for FuseInOutLambdas (#51389)
When doing input fusion, the original input tensor is replaced by a lambda (and the tensor is deleted).
But we still need to pass the KGEN tensor object because the fused kernel needs to take it as a parameter,
but this tensor object is unused because we use the lambdas.
Before, I was replacing it with the input #0 of the source elementwise kernel.
In most cases it works because for most elementwise kernel input type == output type, but it's not always true.
Instead we need to use the CreateEmptyTensorsOp to create a fake tensor (uninitialized) with the right KGEN type.

This PR also fix the lowering of CreateEmptyTensor to work with dynamic shapes: lower to ToTensorFromMGP with a size 1 data buffer with the correct dynamic dims.

MOGG_ORIG_REV_ID: 4b89238322b4e04550d8310f6d633ab8f54800e1