[Kernels][GPU] Adds test for AlibiScoreMod and fixes AlibiScoreMod for on-the-fly alibi calculation. (#52220)
- Fixes `bias` calculation for Alibi.
- Removes `AddFactorMod` from the available `score_mod`s and tests (this
was added as a precursor to Alibi for testing purposes of the MHA
wire-up of score_mod - now that Alibi is in this is not needed anymore).
- Adds `AlibiScoreMod` in the `test_score_mod_mha.mojo` (where
`AddFactorMod` was used before).
- Fixes `scale` calculation on CPU side of `test_score_mod.mojo`.

As a follow-up and depending on performance this will be updated
accordingly.

Closes KERN-624

NN_ORIG_REV_ID: 8a6ed87abe87e0bb22a488f208f13cdd501e911d