[Serving] Server fails if worker fails (#52301)
Checks if the worker process is alive on `aenter` and will terminate if
not so.
Tested with a bad serialized model filename and it stops the server from
coming up. Without this fix, the error is reported, but the server still
comes up.
part of SI-774

---------

Signed-off-by: Rashid Kaleem <rashid@modular.com>
SERVE_ORIG_REV_ID: 194f9fca7c33b2aea7043dcedeeb48744f17c68e