[Kernels] Remove the compile_assembly is_failable feature (#52466)
This removes the is_failable feature from the mojo side. This is a prep
to remove the is_failable from the compiler to simplify the
compile_assembly logic. For historic reasons we added the is_failable to
enable autotuning where the function to be tuned won't be compiled.
Since we no longer drive autotuning through mojo, this feature is no
longer used and can be deleted.

COMPILE_ORIG_REV_ID: 452010449d579fc823dc8c52f4da849ed7ff1bf4