[stdlib][GPU] Fix invalid constant addr space
On Nvidia, address space value of 2 in the IR is reserved.  So, it's
invalid for us to represent the constant address space with the value of
`2`.

Good news: nobody is setting this value currently.  So, just remove it.
While here, rename the existing `PARAM` address space name to be called
"CONSTANT" instead, which is more consistent with the
[docs](https://docs.nvidia.com/cuda/nvvm-ir-spec/#address-space).  Note
that AMD calls address space 4 "constant" as well.

EXTENSIBILITY_ORIG_REV_ID: 35bbc8ad2b82efdddf6fa4cf537132af2ca2d3af