[pipelines] Add kv cache ops for SmolLM and test for generate_llm_logits (#53055)
Two changes:
1. Added opaque cache ops for SmolLM so it can use the continuous cache.
The results appear to be slightly better than with the naive cache with
padding size set to 1 (see [naive output
[Internal link]
	```
Gravity is the force that holds the Earth and the Moon together. It is
the force that keeps the Earth and the Sun in their orbits around the
Sun.

The Moon is a satellite of the Earth. It is a large, round, rocky body
that orbits the Earth. It is about 3,470 miles (5,600 km) in diameter.
It is about 3,470 miles (5,600 km) in diameter. It
	```
2. Added test for `generate_llm_logits` using SmolLM which runs during
presubmit. This lets us get faster error signals if models fail during
the pipeline verification workflow due to API changes.

Closes AIPIPE-277

NN_ORIG_REV_ID: 14c7e5f12034ffe6bd5e81aea01f099f3b89e003