[Extensibility] Implement missing case for mo.conv layer (#53244)
The implementation of mo.conv in the old API was updated to include
a GPU specific case. Without this case, we crash in conv_nhwc_direct
when executing Llama 3.2 on the GPU.

This change fixes up the logic in the extensibility API to include the
relevant logic.

Fixes: GEX-1541
MOGG_ORIG_REV_ID: c9d287bcee75c3da1018a01ae12f73a07bafead9