[SDK] Use subclass ctor in Dim.__new__ (#53339)
Use the subclass's constructor when appropriate in the `Dim.__new__`
special method.
The `__new__` method is used to override class creation and usually
calls the `__init__` method of the class or a subclass.
So for `Dim`, the purpose of this appears to be so that `Dim("x")`
returns a `SymbolicDim`, and `Dim(12)` returns a `StaticDim`.
However, `Dim.__new__` also takes `Dim` subclass arguments but would
previously dispatch to the type of the argument rather than the class in
this case.
For example `StaticDim(SymbolicDim("x"))` would call
`SymbolicDim.__init__`.
So change this to respect the `cls` argument to `__new__`.
This way we hit `StaticDim.__init__` as expected, which fixes an error
message when trying to call `int(Symbolic("x"))`.

Reference: Python data model
[docs](https://docs.python.org/3/reference/datamodel.html).
MAX_GRAPH_API_ORIG_REV_ID: c422931ecf8ab65a37aa9d3a134c4c51656f7614