[stdlib] Combine indexer traits and optimize uint indexing
This sets up removing the Uint -> Int implicit conversion, the majority
of which occur when indexing into collections, pointers, or creating
n-dimensional index types.

It removes the `__index__(self) -> Int` trait which was the same sig as
`__int__(self) -> Int`, and instead uses a single `Indexer` trait for
all indexing.

This enables optimizing away the check to normalize a negative number
when using a UInt, and using SIMD scalar integer types as an index
across a wider variety of collections.

The `Indexer` trait was also used to enable implicit conversions to
`Int` on a few types such as `Bool` and `PythonObject`, this has been
moved to `ImplicitlyIntable` to align with `ImplicitlyBoolable`.

MODULAR_ORIG_COMMIT_REV_ID: 3d5968b3155b4751bd73aaa3d90709ebdfa43aa0