[Serving] Clean up duplicate Dockerfile for Serve (#53564)
For 24.6 we started uploading to Docker Hub, and it was based on the
original Dockerfile (but with a .openai-api appended to name). This PR
gets rid of the duplicate Dockerfiles in preparation for preparing a
"slim" MAX docker container without Pytorch.

EDIT:
[Internal link] to
a run building new image on the branch as a sanity check.
SERVE_ORIG_REV_ID: e873511cc620fedc0ba4bc76ec8a58faab76b35f