[stdlib] Remove `bencher` docs
Now that we have nightly docs, people can see the documentation for
otherwise closed-source things in the standard library, such as the
`benchmark` module. So, remove the temporary markdown docs we provided
as an interim thing for users and contributors of the standard library.
They can be read in a more up-to-date form at
https://docs.modular.com/nightly/mojo/stdlib/benchmark.

MODULAR_ORIG_COMMIT_REV_ID: f5dfec06fc337ca87349888e14079971d09e75b9