[SDK] Add an _add_op variant that returns the op (#53975)
Add a version of `Graph._add_op` that returns the op.
This is needed in a follow up to set attributes on an op after staging.
The current version just returns results, so this makes it easier to
retrieve the op (without doing something like looking at the last block
added to the op, owners of the result values, and so on).

MAX_GRAPH_API_ORIG_REV_ID: ffdb57b3fdb4f86870aa22b99caa17173cde1380