[Kernels] Parameterize `top_k` CPU on `out_idx_type` (#54137)
In prep for merging the CPU and GPU implementations, the GPU
implementation allows specifying the `out_idx_type` even though the CPU
fixes it to be `DType.int64`. Add a parameter for specifying the
`out_idx_type` and fix it to be `DType.int64` for now.

NN_ORIG_REV_ID: 959eeaabe88617da7b48973a8d1c02b2df63e573