[E2E] Make mo.kv_collection_ctor.paged a custom parameteric op (#54214)
This uses custom op params (introduced in #53973) to remove a bunch of
instantiations of mo.kv_collection_ctor.paged. There are
hundreds/thousands of lines that can be removed via this pattern, but
this is a start. This is also important to have more broader support for
other models which just need a new num_heads / head_dim channel to
operate

MOGG_ORIG_REV_ID: f1cce3cadb7215e3fef742125726b6f11123ead3