[SDK] Add an example of _dump_gpu_asm / _dump_gpu_llvm opt use (#54943)
This adds an example (test case) of how to use the dump_gpu_asm or
dump_gpu_llvm option on the session options. These are hidden option
(folks should not reach for them unless they know what they are doing)
which accept either a True/False (to print or not to print) or a Path
which could have a '%' to be replaced by the system by a unique key for
the kernel.

This should enable easy debugging and is compatible with all the
operations that run on GPU within MAX.

MAX_GRAPH_API_ORIG_REV_ID: f7e8ad1bccad3c89770d07ac1dfb2046fa4c2106