[Bindings] Remove private module shims (#55058)
Currently, we define a `core` module for each private module in `max`
that is a compiled extension, then define a "real" private module that
just forwards all the items from `core`.

So the "real" module is (for example) `max._dtype.core`. This layer is
unnecessary, so this PR removes the shim modules and allows them to be
used directly. There are a few things to note about why things are layed
out as they are:
- We need to define them directly in `max`, as (for example) the
compiled extension module defining `_dtype` needs to live directly in
the `max` folder in order for it to be importable as `max._dtype`.
- The `_dtype` (and related) folders need to be removed, as sometimes
the import mechanisms can get confused if there's a `_dtype` folder and
a `_dtype` extension module.

Somewhat merging `modular_nanobind_extension` and
`modular_python_binding_py_library` could be done separately, but I
found it to clean things up by cleaning up boilerplate.

This should also be helpful for making the stubfiles generated directly
from the extensions, without having to hand-maintain them. I have a WIP
PR for that here: [Internal link]

Future TODO/if a PT team member could direct me: Fix the `core` module
in PyTorch. Moving the BUILD file up a directory breaks some bazel
globbing in `//PyTorch:modular_pytorch`, so not really sure what to do
there. Leaving as-is for now.

MAX_GRAPH_API_ORIG_REV_ID: 5c1f5e08ff62a63421fbed36dd34dab989f46073