[Serve] Support none option for tools_choice (#55125)
This PR makes MAX Serve support "none" as an option for `tool_choice`.
If this option is provided, no tool is used and the model generates
output without them. This is even when tools are provided in the
request.

SERVE_ORIG_REV_ID: aa3bf3d51c424760bc99159e2ae4bd64aaa7660b