[Kernels][GPU][Reland] Make the launch bounds parameteric on accelerator (#55170)
This adds a new MAX_THREADS_PER_BLOCK_METADATA that would allow us to
specify the max threads per block for both AMD and NVIDIA GPUs in an
agnostic fashion. This was an optimization that was missing for AMD
hardware because of the lack of generalization.

Reland changes:
* using proper ROCDL attribute to represet flat-work-group-size metadata
* initialization of ROCDL in remaining place of PyTorch

Authored-by: Abdul <adakkak@modular.com>
ALGORITHM_ORIG_REV_ID: 96ce389829668e1e2dab96bb3b27b1aedfd0759b