[Pipelines] Update tokenizer init and enforce encode length (#55262)
Raises an error if the encoded prompt is larger than the max length
defined in the PipelineConfig (or if that's undefined, the max length is
from the HuggingFace's tokenizer_config).

Also, replaces the use of PipelineConfig in the tokenizer with separate
arguments for cleaner code.

Before this change, the tokenizer prints out a warning if the input is
larger than the model's max length, but still returns the full array of
encoded tokens. This causes a problem in MPNet, which uses a positional
embeddings table which hard-caps the sequence length to 514 (since the
position embeddings table's size is [514, hidden_size]). Due to this,
the length of the encoded tokens cannot exceed the model's max length.

SERVE_ORIG_REV_ID: ed8ea86015ca3f87a53344722f802ac6b37b0ad1