[Serving] Schedule request with long-context even if it exc… (#55949)
…eeds the `target_tokens_per_batch_ce` limit.

We have a parameter in our scheduler, `target_tokens_per_batch_ce`,
which controls the number of tokens we aim to include in a batch. It is
default to 4096.

Previously, our logic first checked whether the number of tokens in the
batch exceeded `target_tokens_per_batch_ce`. If it did not, we added a
new request to the batch, regardless of its length. As a result, the
final batch always exceeds `target_tokens_per_batch_ce`.

We need this logic because if we strictly enforce that the final
scheduled batch stays within `target_tokens_per_batch_ce`, we would
never schedule a single request that exceeds this limit.

Closes GENAI-73

SERVE_ORIG_REV_ID: 0169c88308d1ebef9e809fd57a66b3e75e036031