[MOGG] Consolidate handling of DeviceContextPtr for multi-device support (#56325)
Multi-device support currently allows for the extra device arguments to
be passed as multiple `DeviceContextPtr`s or as
a `StaticTuple[DeviceContextPtr]`. The case where the kernel recieves
multiple `DeviceContextPtr` arguments creates ambiguity for kernels when
we convert `MojoCallContextPtr` to `DeviceContextPtr`, as it is
ambiguous whether a `DeviceContextPtr` argument corresponds to the
kernel's device or one of its arguments' devices.

This change removes support for multiple additional `DeviceContextPtr`
arguments, requiring the multi-device `DeviceContextPtr`s to be passed
in a `StaticTuple`.

Towards: GEX-1877
MOGG_ORIG_REV_ID: 379b1a6cd535816d695b7b40dec6fdc49a593925