[mojo-stdlib] Rewrite `IntLiteral` and `FloatLiteral` to use dependent types.
This patch rewrites the `IntLiteral` and `FloatLiteral` types to define
away
bugs due to their long-standing but incorrect design. The problem is
that
the "value" that these types represent are infinite precision literals
that are
only representable at compile time, not materializable at runtime.  The
previous design held this as runtime state, which led to bugs in the
corner
cases, e.g. this would crash the compiler because they can't be
codegen'd:

```mojo
@export
fn test(a: IntLiteral) -> IntLiteral: return a+1
```

We can correct this by building on the recent improvement to dependent
type support (eg parameter inference bugfixes), the movement of all the
related intrinsics to an attribute form, and the
`always_inline(builtin)`
feature being settled, we can correct this by moving these to being
parameters on the literal types instead of runtime values.  In short:

```mojo
# old design
struct IntLiteral:
    var state : __mlir_type.`!kgen.int_literal`
# new design
struct IntLiteral[value: __mlir_type.`!kgen.int_literal`]:
   # .. no state
```

One unfortunate casualty in this is that the special numbers cannot be
accessed as aliases on the type because we don't know the parameters.
As such, this temporarily renames `FloatLiteral.infinity` to
`FloatLiteral_infinity`. If this is a hardship, we can look at improving
the
compiler to move them back.

This also paves the way to remove all the old "kgen.*literal.*` ops, but
that will be a later patch.

NN_ORIG_REV_ID: 237015b553a0fae70aaf294e4f467b9c1312911e