[Serve] Do not schedule long requests twice when chunked prefill is disabled (#56503)
This PR fixes a bug in the Scheduler when we scheduled some requests
twice under some conditions. It only happens when chunked prefill is
disabled and requests length is greater than
`target_tokens_per_batch_ce`.

It was introduced in [Internal link]
The bug is that after we scheduled a long request
(>`target_tokens_per_batch_ce`) it falls through and adds it back to the
input queue.

The benchmarking number don't change. I can't find a dataset and set of
parameters to reliably show the difference. Most probably because that
request has allocated cache slot already, active batch is a dictionary
and the context(data) is mutable and adding it again to the batch won't
change anything. It just leads to inconsistent state.

SERVE_ORIG_REV_ID: 9280d7726149291c1172ee58b078128afdc08d13