[Kernels] [NFC] Replace `__` with `_` in private fns (#56697)
The intent of these functions was to be private and not documented, but
the Mojo doc tooling only respects this for a single `_` prefix, not a
double `__` prefix. Adjust those with `__` to use `_` so they are
properly ignored in the API docs. This cleans up the docs for the
`layout` package a bit as we will start documenting it soon.

Fixes MSTDL-1143.

LAYOUT_ORIG_REV_ID: 39b68ea0a84453deb96b99dae42ba6d389cd8a71