[GPU][docs] Fix module-level doc (#56727)
On the gpu package overview at
https://docs.modular-staging.com/mojo/stdlib/gpu/, the `all_reduce` docs
are showing up empty. This is due to an extra newline at the beginning
of the module level docs. Remove the extra newline so it is rendered
correctly in the package overview.

GPU_ORIG_REV_ID: 207d999667076ebb79f2a9a7a14d54be6b8944c4