[Kernels] [AMD] Add custom scheduling of instructions in matmul  (#56679)
Similar to [Internal link] for the
layout version of matmul. This is about ~506 TFlops for 8192^3 vs ~490
in main. The index version is at around 530, so we have some ways to
cover. But @raiseirql is working on new layout for global loads so I
will look into it once he gives a go ahead.

Also, removing the offset calculations which is currently using ptrs get
this kernel to 514 TFlops, need to figure out better abstraction to
enable that in the future.

LAYOUT_ORIG_REV_ID: 3138fd73bf0c7de5cb42ae40c91119ce797dbd3e