[Pipelines] Remove use of pipeline_config.dtype in architectures
Refactor PipelineModel to accept a SupportedEncoding directly. This
allows us to infer the dtype directly, and avoid leaning on the
PipelineConfig for this information. This is further to support the
seperation of model architectures from the PipelineConfig.

MODULAR_ORIG_COMMIT_REV_ID: 44cd78fdc33c41ed7ce4b71c6486497929f698bd