[Kernels][GPU] Fix layout tensor copy_sram_to_dram bug for non-swizzled mode and masked destination buffer (#57294)
Thanks to @hengjiew for finding this out, this PR fixes a bug in
`copy_sram_to_dram` function. We were using the `copy_from` function
when both source and destination buffers had the same type and it was in
a non-swizzled mode but the destionation buffer was masked. This
resulted in OOB access and overwriting device buffers when `M` and `N`
dimensions were not a power of two.

close KERN-1649

LAYOUT_ORIG_REV_ID: 5dd5a79109238db8709e7825b3a3e1767ad6cc05