[Kernels] [AMD] Fix test_flash_attention_amd.mojo, set entire mask tensor to zero (#57518)
We were not setting the entire mask tensor to zero which was leading to
non-deterministic behavior. The problem was hidden because
`assert_with_measure[cosine]` is very forgiving. It does not throw for
incorrect results if one of the value is 0 or nan. Also the tolerance
were not quite right.

GPU_TEST_ORIG_REV_ID: ff60165fb3e137618fb83b8716f03ce2dd96864a